Rename send_onchain snippets to pay_onchain (#146)
