+++ snippets/dart_snippets/lib/pay_onchain.dart
