+++ snippets/go/pay_onchain.go
