+++ snippets/react-native/pay_onchain.ts
