+++ snippets/rust/src/pay_onchain.rs
